+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 * *[v1.32.0.rc1]*
   * [DESIGN]: Added styles for campaign summary on manage campaign screen
-  * [FEATURE]: Add campaign summary to Manage Campaign screen
+  * [FEATURE]: Add campaign summary to Manage Campaign screen (REEL-67)
+  * [FEATURE]: Add preset Call To Action options (REEL-31)
   * [DESIGN]: Changed website background, text and form input color
 
 ## v1.31.0 (January 8, 2016)